requirements.txt
asgiref==3.10.0
Brotli==1.1.0
cffi==2.0.0
cssselect2==0.8.0
Django==5.2.7
django-crispy-forms==2.4
django-extensions==4.1
django-widget-tweaks==1.5.0
easy-thumbnails==2.10.1
fonttools==4.60.1
mysqlclient==2.2.7
pillow==12.0.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
sqlparse==0.5.3
style==1.1.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
